Backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
PyJWT >= 2.8.0